# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel REST API",
        "description": "Laravel REST API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/product_category/list": {
            "get": {
                "tags": [
                    "Product category"
                ],
                "summary": "Product category list",
                "operationId": "34091dc2180b0cef1616241eb4711981",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter product categories by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Filter product categories by description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product category list"
                    }
                }
            }
        },
        "/api/products/list": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Products list",
                "operationId": "5d999bab0c080922f21c0498b4ce5baa",
                "responses": {
                    "200": {
                        "description": "products list"
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User list",
                "operationId": "27ffe0ec194e07078beae4a3e03bfe28",
                "responses": {
                    "200": {
                        "description": "user list"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Product category",
            "description": "Product category"
        },
        {
            "name": "Products",
            "description": "Products"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}